predictions/mlb/requirements.txt
# predictions/mlb/requirements.txt
# MLB Prediction Worker Dependencies

# Web Framework
flask==3.0.0
gunicorn==21.2.0

# Google Cloud
google-cloud-bigquery==3.13.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-cloud-monitoring==2.17.0

# Pin grpcio and grpcio-status to avoid dependency resolution timeout during builds
grpcio==1.76.0
grpcio-status==1.62.3

# Data Processing
numpy==1.26.2
pandas==2.1.3
db-dtypes==1.2.0  # Required for BigQuery pandas integration

# Machine Learning
xgboost==2.0.2
lightgbm==4.1.0
scikit-learn==1.3.2
catboost==1.2.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0
urllib3==2.6.3  # Pin to match NBA worker — avoids circular import with google-cloud-*

# Testing (optional, for local testing)
pytest==7.4.3
pytest-cov==4.1.0
google-cloud-secret-manager==2.16.0
